// src/openapi/lottery/lottery/models/LotteryItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * lottery.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LotteryItem
 */
export interface LotteryItem {
    /**
     * 
     * @type {string}
     * @memberof LotteryItem
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof LotteryItem
     */
    value?: number;
}

/**
 * Check if a given object implements the LotteryItem interface.
 */
export function instanceOfLotteryItem(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function LotteryItemFromJSON(json: any): LotteryItem {
    return LotteryItemFromJSONTyped(json, false);
}

export function LotteryItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): LotteryItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'value': !exists(json, 'value') ? undefined : json['value'],
    };
}

export function LotteryItemToJSON(value?: LotteryItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'value': value.value,
    };
}