// src/openapi/lottery/lottery/models/LotteryRemark.ts
/* tslint:disable */
/* eslint-disable */
/**
 * lottery.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LotteryRemark
 */
export interface LotteryRemark {
    /**
     * 
     * @type {string}
     * @memberof LotteryRemark
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof LotteryRemark
     */
    require?: boolean;
}

/**
 * Check if a given object implements the LotteryRemark interface.
 */
export function instanceOfLotteryRemark(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function LotteryRemarkFromJSON(json: any): LotteryRemark {
    return LotteryRemarkFromJSONTyped(json, false);
}

export function LotteryRemarkFromJSONTyped(json: any, ignoreDiscriminator: boolean): LotteryRemark {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'require': !exists(json, 'require') ? undefined : json['require'],
    };
}

export function LotteryRemarkToJSON(value?: LotteryRemark | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'require': value.require,
    };
}