// src/openapi/lottery/lottery/models/LotteryType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * lottery.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * - NUMBER: 个数
 *  - PERCENT: 几率
 * @export
 */
export const LotteryType = {
    Number: 'NUMBER',
    Percent: 'PERCENT'
} as const;
export type LotteryType = typeof LotteryType[keyof typeof LotteryType];


export function LotteryTypeFromJSON(json: any): LotteryType {
    return LotteryTypeFromJSONTyped(json, false);
}

export function LotteryTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): LotteryType {
    return json as LotteryType;
}

export function LotteryTypeToJSON(value?: LotteryType | null): any {
    return value as any;
}